EdgeInsets is Double, so a padding can be a fraction of a point
#128. SwiftUI's `EdgeInsets` is `CGFloat`; this one was `Int`, which made
`.padding(0.5)` and any inset derived by division impossible to express.

The change is smaller than it looks, because `computeLayout` was ALREADY
converting to `Double` on the line after reading the insets. What actually
needed deciding was where the rounding goes, and the answer is: once, at the
hand-off to a backend, through the same `Position.vector` every other backend
call already uses. Three places did it differently and are now consistent:

  - `PaddingModifier.commit` built `SIMD2(insets.leading, insets.top)` directly;
    it now builds a `Position` and takes `.vector`, so a fractional inset does
    not invent a second rounding rule.
  - `List.swift` rounded the row height and THEN added the base padding. With
    `Double` insets that order drops the fraction on EVERY row, so it rounds
    once, after adding.
  - `AndroidBackend+List` truncated with `Int(Float(dividerHeightPx) / density)`.
    A 3px divider at density 2.75 is 1.09dp and was reported as 1, losing 8% of
    the divider per row. The fraction now survives.

Nothing at a call site had to change: all 124 `.padding(...)` calls across
Sources/ and testapp/ pass integer literals, and not one passes a variable, so
there was no `Int` binding to break. That was counted, not assumed -- the count
is `grep -rn "\.padding(" Sources/ testapp/*.swift | wc -l`, next to the number
so the next reader can re-derive it.

`defaultPaddingAmount` stays `Int` on the backend protocol. It is a
backend-chosen default in whole points on all five, and widening it would touch
six files to express nothing new.

What is verified and what is not: the whole package plus both Windows backends
build clean, and P61 runs on the rebuilt SwiftCrossUI with its layout intact.
The sub-pixel effect itself is deliberately NOT separately measured, because
after `roundSize` a 0.4pt inset is indistinguishable from a 0pt one on screen --
which is the design, not a gap. What changed is that the value survives to the
rounding instead of being truncated at every step before it.

+++ Sources/AndroidBackend/AndroidBackend+List.swift
@@ -93,7 +93,14 @@ extension AndroidBackend {
 
         let dividerHeightPx = listView.as(AndroidKit.ListView.self)!.getDividerHeight()
 
-        return EdgeInsets(bottom: Int(Float(dividerHeightPx) / density))
+        // No longer truncated to an Int: a divider of 3px at density 2.75 is
+        // 1.09 dp, and the old `Int(...)` reported 1, losing 8% of the divider
+        // on every row. `EdgeInsets` is `Double` now, so the fraction survives
+        // to the single rounding at the backend hand-off.
+        // 不再截斷成 Int:密度 2.75 之下,3px 的分隔線是 1.09 dp,而舊的 `Int(...)`
+        // 會回報 1,於是**每一列**都少掉這條分隔線的 8%。`EdgeInsets` 現在是 `Double`,
+        // 因此那個分數會一路存活到「交給 backend 時的那唯一一次取整」。
+        return EdgeInsets(bottom: Double(Float(dividerHeightPx) / density))
     }
 
     public func minimumRowSize(ofSelectableListView listView: Widget) -> SIMD2<Int> {
